requirements.txt
# Core dependencies
python-dotenv
requests
pandas
openai

# Frontend dependencies
streamlit
plotly

# Additional backend dependencies - remove if not needed
# sqlalchemy